move addresses restaurant relation end point to restaurant route
